+++ angular-workspace/projects/ni/nimble-angular/CHANGELOG.md
@@ -1,8 +1,32 @@
 # Change Log - @ni/nimble-angular
 
-This log was last generated on Thu, 17 Feb 2022 19:43:53 GMT and should not be manually modified.
+This log was last generated on Mon, 21 Feb 2022 23:14:18 GMT and should not be manually modified.
 
 <!-- Start content -->
+
+## 1.1.10
+
+Mon, 21 Feb 2022 23:14:18 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v2.1.3
+
+## 1.1.9
+
+Mon, 21 Feb 2022 15:34:13 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v2.1.2
+
+## 1.1.8
+
+Sat, 19 Feb 2022 00:06:17 GMT
+
+### Patches
+
+- Bump @ni/nimble-components to v2.1.1
 
 ## 1.1.7